Updated draw functions and they now use scoped matrices for displacement
Instead of translate-to and translate-back, it now uses gl::scopedModelMatricies, which create a new 'layer', and is removed when the object goes out of scope.  This way, all rotation/translation is done within the 'layer' and then it is reset when the function exits.  draw::drawStrokedTriangle is also fixed - gl::drawStrokedTriangle was removed so the segments are drawn manually.